2012-10-16 09:36 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * source/ct/atadjust.c   * source/ct/atnum.c   * source/ct/atrepl.c   * source/ct/ctstr.c   * source/ct/ctwfunc.c   * source/ct/ctwin.c   * source/ct/dattime2.c   * source/ct/numat.c   * source/ct/pack.c   * source/ct/print.c   * source/ct/screen1.c   * source/ct/settime.c   * source/ct/dattime3.c   * source/ct/screen2.c     ! more MSVC warnings (W4) suppressed

+++ xharbour/source/ct/ctwfunc.c
@@ -214,7 +214,7 @@ HB_FUNC( WBOX )
       HB_MEMCPY( szBoxBuf, pWBoxFrames[ iFrame ], 9 );
       if( ( iFrame & 4 ) == 0 )
       {
-         szBoxBuf[ 8 ] = hb_gtGetClearChar();
+         szBoxBuf[ 8 ] = ( BYTE ) hb_gtGetClearChar();
       }
       szBoxBuf[ 9 ]  = '0';
       szBox          = szBoxBuf;
